requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
scikit-learn
numpy
